requirements.txt
asgiref==3.8.1
chardet==5.2.0
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Django==5.0.4
django-extra-fields==3.0.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
imageio==2.34.1
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
lazy_loader==0.4
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
regex==2024.4.28
scikit-image==0.23.2
scipy==1.13.0
six==1.16.0
sqlparse==0.5.0
summary==0.2.0
sympy==1.12
tifffile==2024.4.24
torch==2.3.0
torchsummary==1.5.1
torchvision==0.18.0
tqdm==4.66.2
triton==2.3.0
typing_extensions==4.11.0
